forks(n) takes u64 and saturates: the overflow panic dissolves
Party::forks and Clock::forks take n: u64; Forks::new carves
n.saturating_add(1) subregions, so the one input with no headroom for
the residual (n == u64::MAX) yields one share fewer than asked instead
of panicking in debug and wrapping in release. The # Panics sections
pinning the old overflow behavior are gone with the behavior; the
saturating corner is documented at both entry points and on the
iterators.

Split counts in u64; size_hint converts with the standard spelling for
counts past usize ((usize::MAX, None)), documented on Forks. The
forks_max witness suite pins the new total contract in both profiles:
u64::MAX accepted, no panic, the claimed count u64::MAX - 1, shares
genuinely produced and disjoint, and the caller's party intact after
the drop. Callers sweep usize -> u64 (laws, clock tests, fuzzfit
harness and guest).

+++ crates/before/fuzzfit/guest/src/lib.rs
@@ -707,7 +707,7 @@ pub extern "C" fn ff_party_fork(dst: u32, src: u32) -> i32 {
 /// the iterator borrows the source, which keeps its remainder).
 #[no_mangle]
 pub extern "C" fn ff_party_forks(dst: u32, src: u32, n: u32) -> i32 {
-    let shares = with_p_mut(src, |p| p.forks(n as usize).collect::<Vec<_>>());
+    let shares = with_p_mut(src, |p| p.forks(u64::from(n)).collect::<Vec<_>>());
     match shares {
         Some(shares) => {
             for (i, share) in shares.into_iter().enumerate() {
